@@ -20,61 +20,62 @@
     "@yarnpkg/fslib": "2.6.1-rc.5",
     "@yarnpkg/libzip": "2.2.3-rc.5",
     "jsonc-parser": "^3.0.0",
-    "tslib": "^2.0.0"
+    "tslib": "^2.0.0",
+    "@storybook/core-server": "6.5.9"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.2.6",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cdk": "13.2.6",
-    "@angular/common": "13.2.7",
-    "@angular/compiler": "13.2.7",
-    "@angular/compiler-cli": "13.2.7",
-    "@angular/core": "13.2.7",
-    "@angular/forms": "13.2.7",
-    "@angular/platform-browser": "13.2.7",
-    "@angular/platform-browser-dynamic": "13.2.7",
+    "@angular-devkit/build-angular": "14.0.2",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cdk": "14.0.2",
+    "@angular/common": "14.0.2",
+    "@angular/compiler": "14.0.2",
+    "@angular/compiler-cli": "14.0.2",
+    "@angular/core": "14.0.2",
+    "@angular/forms": "14.0.2",
+    "@angular/platform-browser": "14.0.2",
+    "@angular/platform-browser-dynamic": "14.0.2",
     "@babel/core": "7.5.4",
-    "@nrwl/angular": "14.1.4",
-    "@nrwl/cypress": "14.1.4",
-    "@nrwl/eslint-plugin-nx": "14.1.4",
-    "@nrwl/jest": "14.1.4",
-    "@nrwl/js": "14.1.4",
-    "@nrwl/linter": "14.1.4",
-    "@nrwl/node": "14.1.4",
-    "@nrwl/nx-cloud": "14.0.3",
-    "@nrwl/storybook": "14.1.4",
-    "@nrwl/workspace": "14.1.4",
-    "@storybook/addon-essentials": "6.4.19",
+    "@nrwl/angular": "14.3.6",
+    "@nrwl/cypress": "14.3.6",
+    "@nrwl/eslint-plugin-nx": "14.3.6",
+    "@nrwl/jest": "14.3.6",
+    "@nrwl/js": "14.3.6",
+    "@nrwl/linter": "14.3.6",
+    "@nrwl/node": "14.3.6",
+    "@nrwl/nx-cloud": "14.1.2",
+    "@nrwl/storybook": "14.3.6",
+    "@nrwl/workspace": "14.3.6",
+    "@storybook/addon-essentials": "6.5.9",
     "@storybook/addon-knobs": "~6.4.0",
-    "@storybook/angular": "6.4.19",
-    "@storybook/builder-webpack5": "6.4.19",
-    "@storybook/manager-webpack5": "6.4.19",
+    "@storybook/angular": "6.5.9",
+    "@storybook/builder-webpack5": "6.5.9",
+    "@storybook/manager-webpack5": "6.5.9",
     "@types/jest": "27.4.1",
     "@types/node": "14.14.33",
     "@types/universal-analytics": "0.4.5",
     "@types/uuid": "^8.3.4",
     "@types/vscode": "1.50.0",
-    "@typescript-eslint/eslint-plugin": "5.10.2",
-    "@typescript-eslint/parser": "5.10.2",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "@yarnpkg/pnp": "^3.1.1-rc.12",
     "cypress": "6.5.0",
-    "eslint": "8.7.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "ngx-build-plus": "13.0.1",
-    "nx": "14.1.4",
+    "jest-preset-angular": "11.1.2",
+    "ngx-build-plus": "14.0.0",
+    "nx": "14.3.6",
     "ovsx": "^0.3.0",
-    "prettier": "2.6.1",
+    "prettier": "2.7.1",
     "rimraf": "^3.0.2",
     "rxjs": "~7.5.5",
     "standard-version": "^9.1.1",
     "ts-jest": "27.1.4",
     "tslib": "^2.0.0",
     "type-fest": "^2.12.1",
-    "typescript": "4.5.5",
+    "typescript": "4.7.4",
     "universal-analytics": "0.5.3",
     "uuid": "^8.3.2",
     "vsce": "^2.6.3",
