@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   coverageDirectory: '../../coverage/apps/vscode-ui',
 
   setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
@@ -19,5 +19,5 @@ module.exports = {
     '^.+.(ts|mjs|js|html)$': 'jest-preset-angular',
   },
   transformIgnorePatterns: ['node_modules/(?!.*.mjs$)'],
-  preset: '../../jest.preset.ts',
+  preset: '../../jest.preset.js',
 };
