@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   displayName: 'vscode-ui-feature-task-execution-form',
 
   setupFilesAfterEnv: ['<rootDir>/test-setup.ts'],
@@ -20,5 +20,5 @@ module.exports = {
     '^.+.(ts|mjs|js|html)$': 'jest-preset-angular',
   },
   transformIgnorePatterns: ['node_modules/(?!.*.mjs$)'],
-  preset: '../../../jest.preset.ts',
+  preset: '../../../jest.preset.js',
 };
